add feature for a single currency column to the migration
We can have a single currency column. This is useful in case
we have multiple money columnhs and all of them are of the
same currency

the money interger columns are suffixed with _money.
this is needed to reflect on the schema and find the
money columns

also add travis.yml file