add test that the currency will be serialized in a fixed format
currency (single or otherwise) is always a string
and it is in smallcase
because, Money.default_currency.id == "inr".to_sym

previously, single_currency was a Money::Currency object
whereas some_money_currency was a string object

bugfix: also the currency object was being serialized as a yaml
which was a bug. facepalm stupid

this was added so that adding validations are easier
ie. no need to make a case-insensitive check
and to fix the above bug

not very happy. might want to add typecasts